Make and customized PopWindow For Sale.

+++ SellerSimulator/Assets/Scripts/Architecture/SellFrame/ISellFrameSource.cs
@@ -10,6 +10,6 @@ namespace Assets.Scripts.Architecture.WareHouse
     {
 
         Result <List<ModelsSaleFrame>>GetAll();
-        Result<bool> PutOnSale(ulong idBox, int countProduct, int priceSale);
+        Result<bool> PutOnSale(int idBox, int countProduct, int priceSale);
     }
 }
